move event_serie stuff into event_views and event_urls
