+++ webook/arrangement/urls/__init__.py
@@ -7,8 +7,7 @@ from .audience_urls import audience_urls
 from .calendar_urls import calendar_urls
 from .confirmation_urls import confirmation_urls
 from .dashboard_urls import dashboard_urls
-from .event_serie_urls import event_serie_urls
-from .event_urls import event
+from .event_urls import event_urls
 from .insights_urls import insights_urls
 from .location_urls import location_urls
 from .note_urls import note_urls
@@ -41,6 +40,5 @@ urlpatterns.extend(requisition_urls)
 urlpatterns.extend(confirmation_urls)
 urlpatterns.extend(arrangement_type_urls)
 urlpatterns.extend(room_preset_urls)
-urlpatterns.extend(event_serie_urls)
 urlpatterns.extend(analysis_urls)
 urlpatterns.extend(event_urls)
